[Shipping, ShippingBundle, CoreBundle] make the carrier-label translateable
