+++ src/CoreShop/Bundle/ShippingBundle/Form/Type/CarrierTranslationType.php
@@ -14,6 +14,7 @@ namespace CoreShop\Bundle\ShippingBundle\Form\Type;
 
 use CoreShop\Bundle\ResourceBundle\Form\Type\AbstractResourceType;
 use Symfony\Component\Form\Extension\Core\Type\TextareaType;
+use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\FormBuilderInterface;
 
 final class CarrierTranslationType extends AbstractResourceType
@@ -24,6 +25,7 @@ final class CarrierTranslationType extends AbstractResourceType
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
         $builder
+            ->add('label', TextType::class)
             ->add('description', TextareaType::class, [
                 'required' => false,
             ]);
