+++ src/CoreShop/Bundle/ShippingBundle/Resources/config/serializer/Model.Carrier.yml
@@ -10,10 +10,6 @@ CoreShop\Component\Shipping\Model\Carrier:
         name:
             expose: true
             type: string
-            groups: [List, Detailed]
-        label:
-            expose: true
-            type: string
             groups: [Detailed]
         trackingUrl:
             expose: true
