+++ src/CoreShop/Bundle/ShippingBundle/Resources/config/serializer/Model.CarrierTranslation.yml
@@ -2,6 +2,10 @@ CoreShop\Component\Shipping\Model\CarrierTranslation:
     exclusion_policy: ALL
     xml_root_name: carrier_translation
     properties:
+        label:
+            expose: true
+            type: string
+            groups: [List, Detailed]
         description:
             expose: true
             type: string
